pyproject.toml
[tool.black]
line-length = 79

[tool.isort]
line_length = 79
known_local_folder = []

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.poetry]
name = "hf_model"
version = "0.0.1"
description = "..."
authors = ["sentenzo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
torch = { markers = "sys_platform == 'win32'", url = "https://download.pytorch.org/whl/cu118/torch-2.1.2%2Bcu118-cp311-cp311-win_amd64.whl" }
outlines = "^0.0.37"
llama-cpp-python = "^0.2.57"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["slow: marks test as slow (deselect with '-m \"not slow\"')"]
